fix: [남현기] 노래 순서 변경이 제대로 되지 않던 현상 해결 [MUSE-S-100]

+++ src/main/java/me/muse/CrezyBackend/domain/song/controller/form/SongOrderRequestForm.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Getter
 public class SongOrderRequestForm {
     final private Long playlistId;
-    final private List<Long> songIndexList;
+    final private List<Integer> songIndexList;
 }
